+++ backend/src/server.ts
@@ -1,0 +1,12 @@
+import { config } from './config/env'; 
+import { connectDB } from './config/db';
+import app from './app';
+
+const startServer = async () => {
+    await connectDB();  // Connect to DB before starting the app
+    app.listen(config.port, () =>
+        console.log(`Server is running on port ${config.port}`));
+
+  };
+  
+  startServer();
